Moved rewarding-related functionalities into separate file
