+++ contracts/mixnet/src/error.rs
@@ -80,14 +80,9 @@ pub enum ContractError {
         identity: IdentityKey,
         address: Addr,
     },
-    #[error("Overflow error!")]
-    Overflow(#[from] cosmwasm_std::OverflowError),
 
     #[error("We tried to remove more funds then are available in the Reward pool. Wanted to remove {to_remove}, but have only {reward_pool}")]
     OutOfFunds { to_remove: u128, reward_pool: u128 },
-
-    #[error("Invalid ratio")]
-    Ratio(#[from] mixnet_contract::error::MixnetContractError),
 
     #[error("Received invalid rewarding interval nonce. Expected {expected}, received {received}")]
     InvalidRewardingIntervalNonce { received: u32, expected: u32 },
@@ -100,4 +95,7 @@ pub enum ContractError {
 
     #[error("Mixnode {identity} has already been rewarded during the current rewarding interval")]
     MixnodeAlreadyRewarded { identity: IdentityKey },
+
+    #[error("Mixnode's {identity} operator has not been rewarded yet - cannot perform delegator rewarding until that happens")]
+    MixnodeOperatorNotRewarded { identity: IdentityKey },
 }
